+++ CHANGELOG.md
@@ -66,6 +66,7 @@ As of the current development build, Microsoft has finished support for running 
 - Renamed `darkTheme` to `darkMode`.
 - Migrated the Target Provider to TypeScript.
 - Transform the zoom-commands to shortcuts.
+- Move the `loadIcons` function to a more central place in the window registration module.
 
 # 1.8.4